set offline sample as 15

+++ mask2former_video/modeling/matcher.py
@@ -103,7 +103,7 @@ class VideoHungarianMatcher(nn.Module):
         for b in range(bs):
 
             out_prob = outputs["pred_logits"][b].softmax(-1)  # [num_queries, num_classes]
-            tgt_ids = targets[b]["labels"]
+            tgt_ids = targets[b]["labels"].to(torch.int64)
 
             # Compute the classification cost. Contrary to the loss, we don't use the NLL,
             # but approximate it in 1 - proba[target class].
